(refactor): cleaned code
