+++ Sources/PMM_NetworkKit/Core/HTTPMethod.swift
@@ -6,11 +6,12 @@
 //
 
 import Foundation
+
 public enum HTTPMethod: String, Sendable {
-case get = "GET"
-case post = "POST"
-case put = "PUT"
-case patch = "PATCH"
-case delete = "DELETE"
-case head = "HEAD"
+    case get = "GET"
+    case post = "POST"
+    case put = "PUT"
+    case patch = "PATCH"
+    case delete = "DELETE"
+    case head = "HEAD"
 }
